requirements.txt
crewai==0.28.8
langchain-openai==0.1.1
langchain-community==0.0.29
duckduckgo-search==5.2.2
python-dotenv==1.0.1